requirements.txt
bcrypt==4.2.0
colorama==0.4.6
mysql-connector-python==9.0.0
numpy==2.2.4
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
rich==14.0.0
six==1.16.0
tabulate==0.9.0
tzdata==2024.2
tqdm==4.62.3
sentry-sdk==2.25.1